# backend/swagger.yaml
openapi: 3.0.0
info:
  title: Wati Health API
  description: Wati Health Recipe and Medical AI Platform Backend API
  version: 1.0.0
servers:
  - url: http://localhost:5001/api
    description: Local server
  - url: https://api.wati.health/api
    description: Production server
paths:
...
  /auth/register:
    post:
      tags:
        - Auth
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
                - displayName
                - acceptedTerms
              properties:
                email:
                  type: string
                password:
                  type: string
                displayName:
                  type: string
                acceptedTerms:
                  type: boolean
                language:
                  type: string
      responses:
        '201':
          description: User created
        '400':
          description: Validation error
        '409':
          description: Email already in use
  /auth/login:
    post:
      tags:
        - Auth
      summary: User login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                password:
                  type: string
      responses:
        '200':
          description: Logged in successfully
        '401':
          description: Invalid credentials
  /auth/resend-verification:
    post:
      tags:
        - Auth
      summary: Resend verification email
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Verification email sent
  /auth/verify:
    post:
      tags:
        - Auth
      summary: Verify account using token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - token
              properties:
                token:
                  type: string
      responses:
        '200':
          description: Account verified successfully
        '400':
          description: Invalid token
  /auth/forgot-password:
    post:
      tags:
        - Auth
      summary: Request password reset email
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
              properties:
                email:
                  type: string
      responses:
        '200':
          description: Reset email sent (if user exists)
  /auth/reset-password:
    post:
      tags:
        - Auth
      summary: Reset password using token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - token
                - password
              properties:
                token:
                  type: string
                password:
                  type: string
      responses:
        '200':
          description: Password updated
        '400':
          description: Invalid token or password
...
  /auth/me:
    get:
      tags:
        - Auth
      summary: Get current user profile
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Current user info
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
    put:
      tags:
        - Auth
      summary: Update user profile
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Profile updated
    delete:
      tags:
        - Auth
      summary: Delete user account (GDPR)
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Account deleted
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    adminApiKey:
      type: apiKey
      in: header
      name: X-Admin-Key
  schemas:
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
        displayName:
          type: string
        role:
          type: string
          enum: [user, admin, super_admin, health_professional]
        organizations:
          type: array
          nullable: true
          items:
            $ref: '#/components/schemas/Organization'
        isVerified:
          type: boolean
...
    Organization:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        slug:
          type: string
        is_active:
          type: boolean
        settings:
          type: object
        userCount:
          type: integer
        status:
          type: string
          enum: [active, suspended, pending]
...